requirements.txt
# --- Web Framework ---
flask>=3.1.0
flask-cors>=5.0.0
python-dotenv>=1.0.1

# --- AI & Orchestration ---
langchain>=0.3.0
langchain-community>=0.3.0
langchain-google-genai>=2.0.0
langgraph>=0.2.0

# --- Vector Database (Python 3.13 Compatible) ---
faiss-cpu>=1.13.1
numpy>=2.1.0
scikit-learn>=1.6.0

# --- Data Ingestion ---
requests>=2.32.0
beautifulsoup4>=4.12.0
feedparser>=6.0.11

# --- Database ---
sqlalchemy>=2.0.30